Extract `clean-html` directive

Conflicts:
	shell/nuxt/index.js
